rename SvmContext to SvmPayloadContext
