+++ crosshatch/Svm/SvmPayloadContext.ts
@@ -3,8 +3,8 @@ import type { Effect } from "effect"
 
 import type { ChainId } from "../ChainId.ts"
 import type { CreatePayloadError } from "../errors.ts"
-import type { Requirements } from "../Requirements.ts"
 import * as SvmAddress from "./SvmAddress.ts"
+import * as SvmAsset from "./SvmAsset.ts"
 
 export interface SvmLatestBlockhash {
   readonly blockhash: Blockhash
@@ -19,7 +19,8 @@ export interface SvmAssetMetadata {
 export interface SvmPayloadContext {
   readonly getLatestBlockhash: (network: typeof ChainId.Type) => Effect.Effect<SvmLatestBlockhash, CreatePayloadError>
 
-  readonly getAssetMetadata: (
-    requirement: typeof Requirements.Type,
-  ) => Effect.Effect<SvmAssetMetadata, CreatePayloadError>
+  readonly getAssetMetadata: (input: {
+    readonly network: typeof ChainId.Type
+    readonly asset: typeof SvmAsset.SvmAsset.Type
+  }) => Effect.Effect<SvmAssetMetadata, CreatePayloadError>
 }
